+++ solidity/core/contracts/Outbox.sol
@@ -161,7 +161,7 @@ contract Outbox is IOutbox, Version0, MerkleTreeManager, Common {
      * @notice Set contract state to FAILED.
      * @dev Called by the ValidatorManager when fraud is proven.
      */
-    function fail() external onlyValidatorManager {
+    function fail() external override onlyValidatorManager {
         // set contract to FAILED
         state = States.Failed;
         emit Fail();
@@ -177,6 +177,7 @@ contract Outbox is IOutbox, Version0, MerkleTreeManager, Common {
     function isCheckpoint(bytes32 _root, uint256 _index)
         external
         view
+        override
         returns (bool)
     {
         // Checkpoint indices are one-indexed.
